Ft circle ci

+++ server/lib/controllers/articleController.js
@@ -5,7 +5,13 @@ Object.defineProperty(exports, "__esModule", {
 });
 exports.default = void 0;
 
+require("core-js/modules/es.error.cause.js");
+
 var _Article = _interopRequireDefault(require("../models/Article"));
+
+var _Like = _interopRequireDefault(require("../models/Like"));
+
+var _Comment = _interopRequireDefault(require("../models/Comment"));
 
 function _interopRequireDefault(obj) { return obj && obj.__esModule ? obj : { default: obj }; }
 
@@ -17,10 +23,21 @@ articleController.getAll = async (req, res) => {
 };
 
 articleController.post = async (req, res) => {
+  if (!(req.user.membership == "admin" || req.user.email == "smbonimpa2011@gmail.com")) {
+    return res.status(401).send({
+      error: 'Unauthorized action.'
+    });
+  }
+
+  const today = new Date();
   const article = new _Article.default({
     title: req.body.title,
     previewImageURL: req.body.previewImageURL,
-    articleBody: req.body.articleBody
+    articleBody: req.body.articleBody,
+    authorID: req.user._id,
+    date: JSON.stringify(today.toJSON()),
+    subject: "TECH NEWS",
+    readingTime: Math.ceil(req.body.articleBody.split(" ").length * 7.7 / 1000) + " MIN"
   });
   await article.save();
   res.send(article);
@@ -28,10 +45,32 @@ articleController.post = async (req, res) => {
 
 articleController.getOne = async (req, res) => {
   try {
-    const article = await _Article.default.findOne({
+    const singleArticle = {};
+    singleArticle.article = await _Article.default.findOne({
       _id: req.params.id
     });
-    res.send(article);
+
+    if (!singleArticle.article) {
+      throw Error;
+    }
+
+    try {
+      singleArticle.likes = await _Like.default.find({
+        articleID: req.params.id
+      });
+    } catch {
+      singleArticle.likes = [];
+    }
+
+    try {
+      singleArticle.comments = await _Comment.default.find({
+        articleID: req.params.id
+      });
+    } catch {
+      singleArticle.comments = [];
+    }
+
+    res.send(singleArticle);
   } catch {
     res.status(404);
     res.send({
@@ -41,6 +80,12 @@ articleController.getOne = async (req, res) => {
 };
 
 articleController.patch = async (req, res) => {
+  if (!(req.user.membership == "admin" || req.user.email == "smbonimpa2011@gmail.com")) {
+    return res.status(401).send({
+      error: 'Unauthorized action.'
+    });
+  }
+
   try {
     const article = await _Article.default.findOne({
       _id: req.params.id
@@ -56,6 +101,11 @@ articleController.patch = async (req, res) => {
 
     if (req.body.articleBody) {
       article.articleBody = req.body.articleBody;
+      article.readingTime = Math.ceil(req.body.articleBody.split(" ").length * 7.7 / 1000) + " MIN";
+    }
+
+    if (req.body.subject) {
+      article.subject = req.body.subject;
     }
 
     await article.save();
@@ -70,6 +120,12 @@ articleController.patch = async (req, res) => {
 };
 
 articleController.delete = async (req, res) => {
+  if (!(req.user.membership == "admin" || req.user.email == "smbonimpa2011@gmail.com")) {
+    return res.status(401).send({
+      error: 'Unauthorized action.'
+    });
+  }
+
   try {
     await _Article.default.deleteOne({
       _id: req.params.id
