+++ server/lib/controllers/commentController.js
@@ -19,8 +19,8 @@ commentController.getAll = async (req, res) => {
 commentController.post = async (req, res) => {
   const comment = new _Comment.default({
     articleID: req.body.articleID,
-    name: req.body.name,
-    email: req.body.email,
+    userID: req.user._id,
+    name: req.user.name,
     commentBody: req.body.commentBody
   });
   await comment.save();
@@ -32,16 +32,23 @@ commentController.getOne = async (req, res) => {
     const comment = await _Comment.default.find({
       articleID: req.params.id
     });
+    if (comment.length == 0) throw error;
     res.send(comment);
   } catch {
     res.status(404);
     res.send({
-      error: "Comment doesn't exist!"
+      error: "Article's comments not found!"
     });
   }
 };
 
 commentController.delete = async (req, res) => {
+  if (!(req.user.membership == "admin" || req.user.email == "smbonimpa2011@gmail.com")) {
+    return res.status(401).send({
+      error: 'Unauthorized action.'
+    });
+  }
+
   try {
     await _Comment.default.deleteOne({
       _id: req.params.id
