+++ server/src/controllers/articleController.js
@@ -1,4 +1,6 @@
 import Article from "../models/Article";
+import Like from "../models/Like";
+import Comment from "../models/Comment";
 
 const articleController = {};
 
@@ -10,11 +12,19 @@ articleController.getAll = async (req, res) => {
 };
 
 articleController.post = async (req, res) => {
-
+  if(!((req.user.membership == "admin" || req.user.email == "smbonimpa2011@gmail.com" ))){
+    return res.status(401).send({ error:'Unauthorized action.'});
+  }
+  const today = new Date();
   const article = new Article({
     title: req.body.title,
     previewImageURL : req.body.previewImageURL,
-    articleBody: req.body.articleBody
+    articleBody: req.body.articleBody,
+    authorID: req.user._id,
+    date: JSON.stringify(today.toJSON()),
+    subject: "TECH NEWS",
+    readingTime: Math.ceil((req.body.articleBody.split(" ")).length * 7.7 /1000) + " MIN"
+
   });
 
   await article.save();
@@ -23,9 +33,22 @@ articleController.post = async (req, res) => {
 
 articleController.getOne = async (req, res) => {
   try {
-
-    const article = await Article.findOne({ _id: req.params.id });
-    res.send(article);
+    const singleArticle = {};
+    singleArticle.article = await Article.findOne({ _id: req.params.id });
+    if(!singleArticle.article){
+      throw Error;
+    }
+    try {
+      singleArticle.likes = await Like.find({ articleID: req.params.id });
+    } catch {
+      singleArticle.likes = [];  
+    }
+    try {
+      singleArticle.comments = await Comment.find({ articleID: req.params.id });  
+    } catch {  
+      singleArticle.comments = [];  
+    }
+    res.send(singleArticle);
 
   } catch {
 
@@ -36,6 +59,9 @@ articleController.getOne = async (req, res) => {
 };
 
 articleController.patch = async (req, res) => {
+  if(!((req.user.membership == "admin" || req.user.email == "smbonimpa2011@gmail.com" ))){
+    return res.status(401).send({ error:'Unauthorized action.'});
+  }
   try {
     const article = await Article.findOne({ _id: req.params.id });
 
@@ -49,6 +75,10 @@ articleController.patch = async (req, res) => {
 
     if (req.body.articleBody) {
       article.articleBody = req.body.articleBody;
+      article.readingTime = Math.ceil((req.body.articleBody.split(" ")).length * 7.7 /1000) + " MIN";
+    }
+    if (req.body.subject) {
+      article.subject = req.body.subject;
     }
 
     await article.save();
@@ -63,6 +93,9 @@ articleController.patch = async (req, res) => {
 };
 
 articleController.delete = async (req, res) => {
+  if(!((req.user.membership == "admin" || req.user.email == "smbonimpa2011@gmail.com" ))){
+    return res.status(401).send({ error:'Unauthorized action.'});
+  }
   try {
 
     await Article.deleteOne({ _id: req.params.id });
